Fix: searchbar: Removes redux from searchbar, abstracted to custom hook
